0.4.0_1

+++ raffle_server/src/controllers/user.controller.js
@@ -8,8 +8,10 @@ const create_user = catchAsync(async (req, res) => {
   const userId = await userService.create_user(req.body);
   res.status(httpStatus.CREATED).send(userId);
 
-  await get_and_save_first_data(req.body.address, req.body.chain_id);
-  await analyze_first_data(req.body.address, req.body.chain_id);
+  const is_worthy_wallet = await get_and_save_first_data(req.body.address, req.body.chain_id);
+  if(is_worthy_wallet){
+    await analyze_first_data(req.body.address, req.body.chain_id);
+  }
 });
 
 const get_user_by_wallet = catchAsync(async (req, res) => {
@@ -36,23 +38,46 @@ const check_nickname_duplication = catchAsync(async (req, res) => {
 
 
 const get_and_save_first_data = async (address, chain_id) => {
+  let final_coll_set;
+  let has_transfers = true;
+  let is_worthy_wallet = true;
   try {
+    // 소유중인 NFT 콜렉션의 정보를 db에 저장
+    let {has_nft_now, coll_set, slug_set} = await nftService.get_nftcoll_opensea(address, chain_id);
     
-    // 지갑주소의 NFT 콜렉션 db에 저장85
-    const {coll_set, slug_set} = await nftService.get_nftcoll_opensea(address, chain_id);
-    console.log(coll_set) //token_address 반환
-    // console.log(slug_set) //slug 반환
+    if(has_nft_now){
+      // 소유중인 nft 콜렉션의 바닥값 가져옴
+      const fp_total = await nftService.get_nft_fp(slug_set);
+      // 소유중인 NFT 각각의 정보 가져옴  //TODO 이건 사실 병렬이 맞음
+      await nftService.get_nft_moralis(address , chain_id);
+      // 지갑주소의 transfer을 db에 저장
+      let transfer_coll_set = await nftService.get_all_NFT_transfers(address, chain_id, fp_total);
+      // 트랜스퍼 콜렉션 - 지갑주소 nft 콜렉션 = 무중복 세트
+      let coll_set_duplicate_removed = nftService.remove_SetA_from_SetB(coll_set, transfer_coll_set);
+      // db에 무중복 세트가 있는지 확인
+      final_coll_set = await nftService.check_collection_exists(coll_set_duplicate_removed);
+    } else {
+      // 지갑주소의 transfer을 db에 저장
+      let transfer_coll_set = await nftService.get_all_NFT_transfers_no_fp(address, chain_id);
+      // 저장해야할 콜렉션 리스트
+      final_coll_set = transfer_coll_set;
+
+
+      if(final_coll_set.size >1){
+        // 이미 팔아버려서 저장되지 않은 NFT 콜렉션 추가로 검색해 db에 저장
+        await nftService.get_and_save_nftcoll(final_coll_set); 
+      } else {
+        has_transfers = false;
+      }
+      
+      if(!has_nft_now && !has_transfers){
+        is_worthy_wallet = false;
+      }
+
+    }
+
+    return is_worthy_wallet;
     
-    await nftService.get_nft_fp(slug_set);
-    await nftService.get_nft_moralis(address , chain_id);
-    // 지갑주소의 transfer db에 저장
-    let transfer_coll_set = await nftService.get_all_NFT_transfers(address, chain_id);
-    // 트랜스퍼 콜렉션 - 지갑주소 nft 콜렉션 = 무중복 세트
-    let coll_set_duplicate_removed = nftService.remove_SetA_from_SetB(coll_set, transfer_coll_set);
-    // db에 무중복 세트가 있는지 확인
-    const missingAddress = await nftService.check_collection_exists(coll_set_duplicate_removed);
-    // 저장되지 않은 NFT 콜렉션 추가로 검색해 db에 저장
-    await nftService.get_and_save_nftcoll(missingAddress);
     //console.log("FINISHED")
 
 
@@ -70,11 +95,6 @@ const analyze_first_data = async (wallet, chain_id) => {
   }
 };
 
-const test = catchAsync(async (req, res) => {
-  let transfer_coll_set = await nftService.get_all_NFT_transfers('0xfFa914c83D851b9Fe372e4bA6A6E131373AA16ab', 1);
-  res.status(httpStatus.OK).send('open');
-});
-
 
 
 
@@ -84,5 +104,4 @@ module.exports = {
   get_user_by_id,
   update_user,
   check_nickname_duplication,
-  test,
 };
